base: move internal code from libsidbase to libsidinternal
There are now two versions of util.[ch] and common.h. Also rename
base64.[ch] to binary.[ch]. Removed some unneeded header file includes,
and linked libraries.
