+++ .clang-format
@@ -78,7 +78,7 @@ ForEachMacros:
   - list_iterate_items_safe_back
 IncludeBlocks:   Regroup
 IncludeCategories:
-  - Regex:           '^"base/common.h'
+  - Regex:           '^"(base|internal)/common.h'
     Priority:        0
   - Regex:           '^"'
     Priority:        1
