+++ src/daemon/Makefile.am
@@ -22,5 +22,6 @@ sbin_PROGRAMS = sid
 sid_SOURCES = sid.c
 
 sid_LDADD = $(top_builddir)/src/resource/libsidresource.la \
+	    $(top_builddir)/src/internal/libsidinternal.la \
 	    $(top_builddir)/src/base/libsidbase.la \
 	    $(top_builddir)/src/log/libsidlog.la
