+++ src/iface/iface.c
@@ -22,6 +22,7 @@
 #include "base/util.h"
 #include "iface/iface_internal.h"
 
+#include <errno.h>
 #include <stdlib.h>
 #include <string.h>
 #include <sys/mman.h>
