+++ src/include/internal/bitmap.h
