+++ src/include/internal/hash.h
