+++ src/tools/sidctl/Makefile.am
@@ -21,7 +21,8 @@ sbin_PROGRAMS = sidctl
 
 sidctl_SOURCES = sidctl.c
 
-sidctl_LDADD = $(top_builddir)/src/base/libsidbase.la \
+sidctl_LDADD = $(top_builddir)/src/internal/libsidinternal.la \
+	       $(top_builddir)/src/base/libsidbase.la \
 	       $(top_builddir)/src/iface/libsidiface.la \
 	       $(top_builddir)/src/log/libsidlog.la \
 	       $(top_builddir)/src/resource/libsidresource.la
