+++ src/tools/usid/usid.c
@@ -17,8 +17,6 @@
  * along with SID.  If not, see <http://www.gnu.org/licenses/>.
  */
 
-#include "base/common.h"
-
 #include "base/buffer.h"
 #include "base/util.h"
 #include "iface/iface.h"
