+++ tests/Makefile.am
@@ -12,9 +12,11 @@ check_PROGRAMS = \
 
 TESTS = $(check_PROGRAMS)
 test_buffer_SOURCES = test_buffer.c
-test_buffer_LDADD = $(top_builddir)/src/base/libsidbase.la -lcmocka
+test_buffer_LDADD = $(top_builddir)/src/internal/libsidinternal.la \
+		    $(top_builddir)/src/base/libsidbase.la -lcmocka
 test_hash_SOURCES = test_hash.c
-test_hash_LDADD = $(top_builddir)/src/base/libsidbase.la -lcmocka
+test_hash_LDADD = $(top_builddir)/src/internal/libsidinternal.la \
+		  $(top_builddir)/src/base/libsidbase.la -lcmocka
 test_kv_store_SOURCES = test_kv_store.c
 test_kv_store_LDADD = \
 	$(top_builddir)/src/base/libsidbase.la \
@@ -24,7 +26,8 @@ test_notify_LDFLAGS = -Wl,--wrap=sd_notify -Wl,--wrap=sid_buffer_get_data
 test_notify_LDADD = \
 	$(top_builddir)/src/iface/libsidiface_servicelink.la -lcmocka
 test_bitmap_SOURCES = test_bitmap.c
-test_bitmap_LDADD = $(top_builddir)/src/base/libsidbase.la -lcmocka
+test_bitmap_LDADD = $(top_builddir)/src/internal/libsidinternal.la \
+		    $(top_builddir)/src/base/libsidbase.la -lcmocka
 test_iface_SOURCES = test_iface.c
 test_iface_LDFLAGS = -Wl,--wrap=read -Wl,--wrap=close -Wl,--wrap=getenv \
 	-Wl,--wrap=sid_comms_unix_init -Wl,--wrap=sid_comms_unix_recv \
